back/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-magic==0.4.27
pdfplumber==0.10.3
openai==1.12.0
python-dotenv==1.0.0